파일 경로 변경 및 검토
