+++ Week2/problem/B1212.java
@@ -7,4 +7,3 @@ public class B1212 {
     // 가능하면 500ms 이내로 풀어봅시다.
     }
 }
-
